refactor(repository): 纯移动拆分 usage_log_repo.go（4701→212行）
